v1.3.0

Conflicts:
	contrib/syntax/textmate/Docker.tmbundle/Syntaxes/Dockerfile.tmLanguage
	contrib/syntax/vim/syntax/dockerfile.vim
	integration/buildfile_test.go
	server/buildfile.go

Docker-DCO-1.1-Signed-off-by: Anthony Baire <Anthony.Baire@irisa.fr> (github: a-ba)

+++ hack/vendor.sh
@@ -39,25 +39,33 @@ clone() {
 	echo done
 }
 
-clone git github.com/kr/pty 98c7b80083
+clone git github.com/kr/pty 67e2db24c8
 
-clone git github.com/gorilla/context 708054d61e5
+clone git github.com/gorilla/context 14f550f51a
 
-clone git github.com/gorilla/mux 9b36453141c
+clone git github.com/gorilla/mux 136d54f81f
 
-clone git github.com/syndtr/gocapability 3454319be2
+clone git github.com/tchap/go-patricia v1.0.1
 
 clone hg code.google.com/p/go.net 84a4013f96e0
 
 clone hg code.google.com/p/gosqlite 74691fb6f837
 
-# get Go tip's archive/tar, for xattr support
-# TODO after Go 1.3 drops, bump our minimum supported version and drop this vendored dep
-clone hg code.google.com/p/go a15f344a9efa
-mv src/code.google.com/p/go/src/pkg/archive/tar tmp-tar
-rm -rf src/code.google.com/p/go
-mkdir -p src/code.google.com/p/go/src/pkg/archive
-mv tmp-tar src/code.google.com/p/go/src/pkg/archive/tar
+clone git github.com/docker/libtrust d273ef2565ca
 
-clone git github.com/godbus/dbus v1
-clone git github.com/coreos/go-systemd v1
+# get Go tip's archive/tar, for xattr support and improved performance
+# TODO after Go 1.4 drops, bump our minimum supported version and drop this vendored dep
+if [ "$1" = '--go' ]; then
+	# Go takes forever and a half to clone, so we only redownload it when explicitly requested via the "--go" flag to this script.
+	clone hg code.google.com/p/go 1b17b3426e3c
+	mv src/code.google.com/p/go/src/pkg/archive/tar tmp-tar
+	rm -rf src/code.google.com/p/go
+	mkdir -p src/code.google.com/p/go/src/pkg/archive
+	mv tmp-tar src/code.google.com/p/go/src/pkg/archive/tar
+fi
+
+clone git github.com/docker/libcontainer 8d1d0ba38a7348c5cfdc05aea3be34d75aadc8de
+# see src/github.com/docker/libcontainer/update-vendor.sh which is the "source of truth" for libcontainer deps (just like this file)
+rm -rf src/github.com/docker/libcontainer/vendor
+eval "$(grep '^clone ' src/github.com/docker/libcontainer/update-vendor.sh | grep -v 'github.com/codegangsta/cli')"
+# we exclude "github.com/codegangsta/cli" here because it's only needed for "nsinit", which Docker doesn't include
